requirements.txt
selene>=2.0.0b4
pytest
allure-pytest==2.9.45
